pyproject.toml
[project]
name = "xngin"
version = "0.0.1"
dependencies = [
    "boto3>=1.35,<1.36",
    "deepdiff>=8.0.1,<8.1",
    "fastapi[standard]>=0.115.6,<0.116.0",
    "gspread>=6.1.2,<6.2.0",
    "httpx>=0.27.0,<0.28.0",
    "numpy>=1.26.4,<1.27",
    "pandas>=2.2.0,<3",
    "pandas-gbq>=0.25.0,<0.26",
    "parsy>=2.1,<3",
    "psycopg>=3.2.3,<4",
    "psycopg2>=2.9.10,<3",
    "pydantic>=2.10.4,<3.0.0",
    "scipy>=1.12.0,<2",
    "sqlalchemy-bigquery>=1.12.0,<2",
    "sqlalchemy==2.0.32",
    "statsmodels>=0.14.1,<0.20",
    "stochatreat>=0.0.20,<0.1.0",
    "typer>=0.12.5,<0.13.0",
    "uvicorn[standard]>=0.32.0",
    "zstandard>=0.23.0,<0.24",
    "sentry-sdk[fastapi]>=2.19.2,<2.20",
    "python-jose>=3.3.0",
]
requires-python = ">=3.12,<3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
xngin-apiserver = "xngin.apiserver.main:main"
xngin-cli = "xngin.cli.main:app"

[tool.pytest.ini_options]
log_cli = 0  # set to 1 for live logs
log_cli_level = "warn"
markers = [
    "integration: integration tests that rely on remote resources",
    "nondeterministic: tests that are nondeterministic (but should be)"
    ]
addopts = "-m 'not integration and not nondeterministic'"
filterwarnings = [  # format is: action:message:category:module
    "ignore:`example` has been deprecated:DeprecationWarning:xngin.apiserver.dependencies",
]
env = [
    "SENTRY_DSX="  # Disable Sentry in unit tests.
]

[tool.ruff]
preview = true
exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "A",
    "ASYNC",
    "B",
    "E",
    "W",
    "F",
    "N",
    "PD",
    "FAST",
    "RET",
    "RUF",
    "SIM",
    "TRY",
    "UP"
]
ignore = [
    "RUF029",
    "TRY003",
    "E501", # redundant line length warnings
    "E266", # Too many leading `#` before block comment
    "PD003", # `.isna` is preferred to `.isnull`; functionality is equivalent"
    "PD002", # `inplace=True` should be avoided; it has inconsistent behavior
    "PD101", # Using `series.nunique()` for checking that a series is constant is inefficient
    "PD901", # Avoid using the generic variable name `df` for DataFrames
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1,<4.1.0",
    "pytest>=8.3.2,<9.0.0",
    "ruff>=0.8.0,<0.9.0",
    "pytest-mock>=3.14.0,<4",
    "types-psycopg2>=2.9.10,<3",
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.3.241126",
    "pytest-env>=1.1.5,<1.2",
]
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]

[tool.mypy]
python_version = 3.12
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "deepdiff"
ignore_missing_imports = true
